requirements.txt
django~=4.2.7
djangorestframework~=3.14.0
django-filter~=23.5
psycopg2-binary

pytest~=7.4.4